Profile deletion

+++ frontend/src/app/navbar/navbar.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
+import { AccommodationsService } from '../pages/accommodations/accommodations.service';
+import { AccommodationDTO } from '../pages/accommodations/model/accommodationDTO';
 import { AuthService } from '../pages/login/log-auth.service';
+import { Reservation, ReservationService } from '../pages/reservations/services/reservation.service';
 import { DeleteDTO } from '../pages/users/deleteDTO';
 import { UserService } from '../pages/users/users.service';
 
@@ -18,7 +21,9 @@ export class NavbarComponent implements OnInit {
   role?: string;
   del: DeleteDTO = {} as DeleteDTO;
   
-  constructor(private authService: AuthService, private router: Router, private userService: UserService) {
+  constructor(private authService: AuthService, private router: Router, 
+    private userService: UserService, private accomService: AccommodationsService,
+    private resService: ReservationService) {
   }
 
   ngOnInit(): void {
@@ -45,10 +50,49 @@ export class NavbarComponent implements OnInit {
 
   delete() {
     let userId = this.authService.getUserId();
-    this.del.id = userId
-    this.userService.delete(this.del).subscribe()
-    this.authService.logout();
-    this.router.navigate(['login']);
+
+    if (this.role === 'Guest') {
+      this.userService.hasActiveReservations(userId).subscribe((res: Reservation[]) => {
+        console.log(res);
+        if (res.length != 0) {
+          alert('Can\'t delete profile because you have active reservations');
+          return;
+        }
+  
+        this.del.id = userId
+        this.userService.delete(this.del).subscribe()
+        this.authService.logout();
+        this.router.navigate(['login']);
+      });
+    } else if (this.role === 'Host') {
+      this.accomService.getAccommodationsByHost(userId).subscribe((res: AccommodationDTO[]) => {
+        let accommodations: AccommodationDTO[] = [];
+        for (let i = 0; i < Object.values(res).length; ++i) {
+          accommodations.push(Object.values(res)[i])
+        }
+
+        if (accommodations.length != 0) {
+          for (let i = 0; i < accommodations.length; ++i) {
+            this.resService.getReservationsByAccommodation(accommodations[i].id).subscribe((res: Reservation[]) => {
+              
+              if (res.length != 0) {
+                alert('Can\'t delete profile because you have active reservations for accommodation with id ' + accommodations[i].id);
+                return;
+              }
+
+            });
+          }
+  
+          console.log(accommodations[0]);
+        }
+
+
+      });
+    }
+
+
   }
+
+
 
 }
